requirements/dev.txt
django==5.0.4
loguru==0.7.3
djangorestframework==3.15.1
drf-yasg==1.21.10
pytest-django==4.10.0
factory-boy==3.3.3
pytest-factoryboy==2.7.0
pytest-cov==6.0.0